refactor(#9): update db query functions location
